+++ app/Http/Controllers/AdminController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\User;
+use App\Models\User;
 use Auth;
 use Illuminate\Http\Request;
