+++ app/Http/Controllers/Auth/RegisterController.php
@@ -4,9 +4,9 @@ namespace App\Http\Controllers\Auth;
 
 use App\Http\Controllers\Controller;
 use App\Mail\Welcome;
-use App\Player;
-use App\Team;
-use App\User;
+use App\Models\Player;
+use App\Models\Team;
+use App\Models\User;
 use Illuminate\Foundation\Auth\RegistersUsers;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Validator;
