+++ app/Http/Controllers/PlayerController.php
@@ -2,10 +2,10 @@
 
 namespace App\Http\Controllers;
 
-use App\Match;
-use App\Player;
-use App\Team;
-use App\User;
+use App\Models\Match;
+use App\Models\Player;
+use App\Models\Team;
+use App\Models\User;
 use Auth;
 use Illuminate\Http\Request;
 use Image;
