+++ app/Mail/nyttMedlem.php
@@ -2,7 +2,7 @@
 
 namespace App\Mail;
 
-use App\Medlem;
+use App\Models\Medlem;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Mail\Mailable;
