+++ app/Models/About.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
-class Index extends Model
+class About extends Model
 {
 }
