+++ app/Models/Game.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
@@ -8,6 +8,6 @@ class Game extends Model
 {
     public function team()
     {
-        return $this->hasMany(\App\Team::class);
+        return $this->hasMany(\App\Models\Team::class);
     }
 }
