+++ app/Models/TournamentMatch.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
@@ -8,11 +8,11 @@ class TournamentMatch extends Model
 {
     public function tournament()
     {
-        return $this->hasOne(\App\Tournament::class);
+        return $this->hasOne(\App\Models\Tournament::class);
     }
 
     public function team()
     {
-        return $this->belongsTo(\App\Team::class);
+        return $this->belongsTo(\App\Models\Team::class);
     }
 }
